@@ -1,9 +1,8 @@
-﻿using Confluent.Kafka;
-
-namespace Reactive.Kafka.Interfaces
+﻿namespace Reactive.Kafka.Interfaces
 {
     public interface IConsumerWrapper
     {
         IConsumer<string, string> Consumer { get; }
+        DateTime LastConsume { get; }
     }
 }
