@@ -1,67 +1,64 @@
-﻿using System.Collections.Generic;
-using Reactive.Kafka.Helpers;
-using Reactive.Kafka.Tests.Types;
-using Xunit;
+﻿using static Reactive.Kafka.Helpers.Convert;
 
 namespace Reactive.Kafka.Tests;
 
 public class HelperConvertTest
 {
     [Fact]
-    public void  TryChangeTypeIsTrue()
+    public void TryChangeTypeIsTrue()
     {
         // Arrange
         const int value = 2;
         const string expected = "2";
-        
+
         // Act
-        var isChange = Convert.TryChangeType(value, out string output);
-        
+        var isChange = TryChangeType(value, out string output);
+
         // Assert
         Assert.True(isChange);
         Assert.True(output == expected);
     }
-    
+
     [Fact]
-    public void  TryChangeTypeIsFalse()
+    public void TryChangeTypeIsFalse()
     {
         // Arrange
         List<int> value = new();
         const int expected = 0;
-        
+
         // Act
-        var isChange = Convert.TryChangeType(value, out int output);
-        
+        var isChange = TryChangeType(value, out int output);
+
         // Assert
         Assert.False(isChange);
         Assert.True(output == expected);
     }
-    
+
     [Fact]
-    public void  TrySerializeTypeIsTrue()
+    public void TrySerializeTypeIsTrue()
     {
         // Arrange 
         const string str = "{\"id\":0, \"name\": \"Kafka\"}";
         const string expected = "MessageTest { Id = 0, Name = Kafka }";
-        
+
         // Act
-        var result = Convert.TrySerializeType(str, out MessageTest output);
-        
+        var result = TrySerializeType(str, out MessageTest output);
+
         // Assert
         Assert.True(result);
         Assert.True(output.ToString() == expected);
 
     }
-    
+
     [Fact]
-    public void  TrySerializeTypeIsFalse()
+    public void TrySerializeTypeIsFalse()
     {
         // Arrange 
         const string str = "valor";
 
         // Act
-        var result = Convert.TrySerializeType(str, out MessageTest output);
-        
+        var result = TrySerializeType(str, out MessageTest output);
+
         // Assert
         Assert.False(result);
         Assert.Null(output);
